+++ src/chalk.nim
@@ -7,7 +7,7 @@
 
 # Note that imports cause topics and plugins to register.
 {.warning[UnusedImport]: off.}
-import "."/[config, confload, commands, norecurse, sinks, docker_base,
+import "."/[config, confload, commands, norecurse, sinks,
             attestation_api, util]
 
 when isMainModule: