+++ CHANGELOG.md
@@ -12,6 +12,13 @@
   it defaulted to `us-east-1` if the `AWS_REGION` or
   `AWS_DEFAULT_REGION` environment variables were not set
   [#246](https://github.com/crashappsec/chalk/pull/246)
+- Removed chalk keys:
+  - `_IMAGE_VIRTUAL_SIZE` - deprecated by docker
+  - `_IMAGE_LAST_TAG_TIME` - scoped to local daemon and is
+    not shared with buildx. Many images report as
+    `0001-01-01T00:00:00Z`
+  - `_IMAGE_STORAGE_METADATA` - metadata of a docker storage
+    driver and is not directly related to docker image
 
 ### Fixes